[BugFix] :    상대경로 절대경로로 변경 (서린-17:38)
