+++ src/utils/getPbImage.ts
@@ -1,5 +1,5 @@
 export function getPbImageURL(item, fileName = 'photo'): string {
-  const pbUrl = import.meta.env.VITE_PB_URL;
+  const pbUrl = import.meta.env.VITE_PB_API;
 
   return `${pbUrl}/api/files/${item.collectionId}/${item.id}/${item[fileName]}`;
 }
